[INTERNAL][CORE] Split off resource negotiation factory
Splits of the ResourceNegotiationFactory from NegotiationFactory.

Moves the ResourceNegotiationFactory into the session context.
