+++ core/src/saros/session/SarosSessionManager.java
@@ -44,6 +44,7 @@ import saros.negotiation.OutgoingSessionNegotiation;
 import saros.negotiation.ResourceNegotiation;
 import saros.negotiation.ResourceNegotiationCollector;
 import saros.negotiation.ResourceNegotiationData;
+import saros.negotiation.ResourceNegotiationFactory;
 import saros.negotiation.ResourceSharingData;
 import saros.negotiation.SessionNegotiation;
 import saros.negotiation.hooks.ISessionNegotiationHook;
@@ -89,6 +90,7 @@ public class SarosSessionManager implements ISarosSessionManager {
   private static final long NEGOTIATION_TIMEOUT = 10000L;
 
   private volatile SarosSession session;
+  private volatile ResourceNegotiationFactory resourceNegotiationFactory;
 
   private final IContainerContext context;
 
@@ -266,6 +268,8 @@ public class SarosSessionManager implements ISarosSessionManager {
       session.start();
       sessionStarted(session);
 
+      resourceNegotiationFactory = session.getComponent(ResourceNegotiationFactory.class);
+
       for (IReferencePoint referencePoint : referencePoints) {
         String referencePointId = String.valueOf(SESSION_ID_GENERATOR.nextInt(Integer.MAX_VALUE));
 
@@ -289,6 +293,8 @@ public class SarosSessionManager implements ISarosSessionManager {
     JID localUserJID = connectionHandler.getLocalJID();
 
     session = new SarosSession(id, localUserJID, host, localProperties, hostProperties, context);
+
+    resourceNegotiationFactory = session.getComponent(ResourceNegotiationFactory.class);
 
     log.info("joined uninitialized Saros session");
 
@@ -346,6 +352,7 @@ public class SarosSessionManager implements ISarosSessionManager {
 
       ISarosSession currentSession = session;
       session = null;
+      resourceNegotiationFactory = null;
 
       sessionEnded(currentSession, reason);
 
@@ -437,9 +444,17 @@ public class SarosSessionManager implements ISarosSessionManager {
         return;
       }
 
+      ResourceNegotiationFactory currentResourceNegotiationFactory = resourceNegotiationFactory;
+
+      if (currentResourceNegotiationFactory == null) {
+        log.warn("could not accept resource negotiation as no session is running");
+
+        return;
+      }
+
       try {
         negotiation =
-            negotiationFactory.newIncomingResourceNegotiation(
+            currentResourceNegotiationFactory.newIncomingResourceNegotiation(
                 remoteAddress, negotiationID, resourceNegotiationData, this, session);
 
         negotiation.setNegotiationListener(negotiationListener);
@@ -631,6 +646,14 @@ public class SarosSessionManager implements ISarosSessionManager {
       return;
     }
 
+    ResourceNegotiationFactory currentResourceNegotiationFactory = resourceNegotiationFactory;
+
+    if (currentResourceNegotiationFactory == null) {
+      log.warn("could not start a resource negotiation as no session is running");
+
+      return;
+    }
+
     List<User> recipients = new ArrayList<>();
     if (session.isHost()) {
       /*
@@ -654,7 +677,7 @@ public class SarosSessionManager implements ISarosSessionManager {
     try {
       for (User user : recipients) {
         AbstractOutgoingResourceNegotiation negotiation =
-            negotiationFactory.newOutgoingResourceNegotiation(
+            currentResourceNegotiationFactory.newOutgoingResourceNegotiation(
                 user.getJID(), resourceSharingData, this, session);
 
         negotiation.setNegotiationListener(negotiationListener);
@@ -673,8 +696,9 @@ public class SarosSessionManager implements ISarosSessionManager {
   public void startSharingReferencePoints(JID user) {
 
     ISarosSession currentSession = session;
+    ResourceNegotiationFactory currentResourceNegotiationFactory = resourceNegotiationFactory;
 
-    if (currentSession == null) {
+    if (currentSession == null || currentResourceNegotiationFactory == null) {
       /*
        * as this currently only called by the OutgoingSessionNegotiation
        * job just silently return
@@ -718,7 +742,7 @@ public class SarosSessionManager implements ISarosSessionManager {
         }
 
         negotiation =
-            negotiationFactory.newOutgoingResourceNegotiation(
+            currentResourceNegotiationFactory.newOutgoingResourceNegotiation(
                 user, currentSharedReferencePoints, this, currentSession);
 
         negotiation.setNegotiationListener(negotiationListener);
