+++ eclipse/test/junit/saros/session/internal/SarosSessionTest.java
@@ -43,6 +43,7 @@ import saros.context.IContextKeyBindings;
 import saros.context.SarosEclipseSessionContextFactory;
 import saros.editor.EditorManager;
 import saros.filesystem.IPathFactory;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.net.IConnectionManager;
 import saros.net.IReceiver;
 import saros.net.ITransmitter;
@@ -267,6 +268,9 @@ public class SarosSessionTest {
 
     addMockedComponent(IPathFactory.class);
     addMockedComponent(ISarosSessionManager.class);
+
+    addMockedComponent(IChecksumCache.class);
+    addMockedComponent(saros.filesystem.IWorkspace.class);
 
     container.start();
   }
